+++ src/main/java/de/gematik/vau/lib/crypto/KEM.java
@@ -16,34 +16,34 @@
 
 package de.gematik.vau.lib.crypto;
 
+
 import de.gematik.vau.lib.data.*;
-
+import java.security.GeneralSecurityException;
+import java.security.KeyPair;
+import java.security.PublicKey;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
-import lombok.SneakyThrows;
+import javax.crypto.Cipher;
+import javax.crypto.SecretKey;
+import javax.crypto.spec.GCMParameterSpec;
+import javax.crypto.spec.SecretKeySpec;
+import lombok.NoArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ArrayUtils;
-
 import org.bouncycastle.crypto.Digest;
 import org.bouncycastle.crypto.digests.SHA256Digest;
 import org.bouncycastle.crypto.generators.HKDFBytesGenerator;
-import org.bouncycastle.crypto.params.*;
+import org.bouncycastle.crypto.params.HKDFParameters;
 import org.bouncycastle.jcajce.SecretKeyWithEncapsulation;
 import org.bouncycastle.jce.interfaces.ECPrivateKey;
 import org.bouncycastle.jce.interfaces.ECPublicKey;
 
-import javax.crypto.*;
-import javax.crypto.spec.GCMParameterSpec;
-import javax.crypto.spec.SecretKeySpec;
-import java.security.*;
-import java.util.ArrayList;
-import java.util.List;
-
+@Slf4j
+@NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
 public class KEM {
   private static final int GCM_IV_LENGTH = 12; //A_24628
   private static final int GCM_TAG_LENGTH = 16; //A_24628
-
-  private KEM() {
-    // Prevent instantiation by making the constructor private
-  }
 
   /**
    * Generates the Shared secrets using own Private and the remote ciphertexts of VauMessage2
@@ -51,7 +51,6 @@ public class KEM {
    * @param privateKeys own PrivateKey
    * @return The ECDH and Kyber Shared Secrets
    */
-  @SneakyThrows
   public static KdfMessage decapsulateMessages(VauMessage2 ciphertext, EccKyberKeyPair privateKeys) {
     ECPublicKey ecdhPublicKeySender = cborDecodeEcdhPublicKey(ciphertext.getEcdhCt());
     byte[] ecdhSharedSecret = EllipticCurve.getSharedSecret(ecdhPublicKeySender, privateKeys.getEccKeyPair()
@@ -59,6 +58,7 @@ public class KEM {
 
     byte[] sharedSecretClient = KyberEncoding.pqcGenerateDecryptionKey(privateKeys.getKyberKeyPair().getPrivate(),
       ciphertext.getKyberCt());
+
     return new KdfMessage(null, ecdhSharedSecret, null, sharedSecretClient);
   }
 
@@ -83,7 +83,6 @@ public class KEM {
    * @param kyberPublicKey remote Kyber PublicKey
    * @return KdfMessage containing both secrets and ciphertexts
    */
-  @SneakyThrows
   public static KdfMessage encapsulateMessage(ECPublicKey remoteEcdhPublicKey, PublicKey kyberPublicKey) {
     KeyPair temporaryEcdhKeyPair = EllipticCurve.generateKeyPair();
     byte[] ecdhSharedSecret = EllipticCurve.getSharedSecret(remoteEcdhPublicKey,
@@ -128,12 +127,16 @@ public class KEM {
       throw new IllegalArgumentException("Kdf Message 1 was null.");
     } else if (message1.getEcdhSharedSecret() == null) {
       throw new IllegalArgumentException("Ecdh Shared Secret of Message 1 was null.");
+    } else if (message1.getEcdhSharedSecret().length != 32) {
+      throw new IllegalArgumentException("Length of Ecdh Shared Secret of Message 1 must be 32.");
     } else if (message1.getKyberSharedSecret() == null) {
       throw new IllegalArgumentException("Kyber Shared Secret of Message 1 was null.");
     } else if (message2 == null) {
       throw new IllegalArgumentException("Kdf Message 2 was null.");
     } else if (message2.getEcdhSharedSecret() == null) {
       throw new IllegalArgumentException("Ecdh Shared Secret of Message 2 was null.");
+    } else if (message2.getEcdhSharedSecret().length != 32) {
+      throw new IllegalArgumentException("Length of Ecdh Shared Secret of Message 2 must be 32.");
     } else if (message2.getKyberSharedSecret() == null) {
       throw new IllegalArgumentException("Kyber Shared Secret of Message 2 was null.");
     }
@@ -167,21 +170,24 @@ public class KEM {
    * @param plaintext text to be encrypted
    * @return resulting ciphertext
    */
-  @SneakyThrows
   public static byte[] encryptAead(byte[] key, byte[] plaintext) {
-    if (key.length != 32) {
-      throw new IllegalArgumentException("Key length must be 32 bytes");
+    try {
+      if (key.length != 32) {
+        throw new IllegalArgumentException("Key length must be 32 bytes");
+      }
+      byte[] iv = new byte[GCM_IV_LENGTH];
+      ThreadLocalRandom.current().nextBytes(iv);
+      GCMParameterSpec ivParameterSpec = new GCMParameterSpec(GCM_TAG_LENGTH * Byte.SIZE, iv);
+
+      SecretKey secretKey = new SecretKeySpec(key, "AES");
+      Cipher cipher = Cipher.getInstance("AES/GCM/NoPadding");
+      cipher.init(Cipher.ENCRYPT_MODE, secretKey, ivParameterSpec);
+
+      byte[] ciphertext = cipher.doFinal(plaintext);
+      return ArrayUtils.addAll(iv, ciphertext);
+    } catch (GeneralSecurityException e) {
+      throw new IllegalArgumentException("Error while encrypting plaintext", e);
     }
-    byte[] iv = new byte[GCM_IV_LENGTH];
-    ThreadLocalRandom.current().nextBytes(iv);
-    GCMParameterSpec ivParameterSpec = new GCMParameterSpec(GCM_TAG_LENGTH * Byte.SIZE, iv);
-
-    SecretKey secretKey = new SecretKeySpec(key, "AES");
-    Cipher cipher = Cipher.getInstance("AES/GCM/NoPadding");
-    cipher.init(Cipher.ENCRYPT_MODE, secretKey, ivParameterSpec);
-
-    byte[] ciphertext = cipher.doFinal(plaintext);
-    return ArrayUtils.addAll(iv, ciphertext);
   }
 
   /**
@@ -190,20 +196,23 @@ public class KEM {
    * @param cipherText ciphertext to be decrypted
    * @return the resulting plaintext
    */
-  @SneakyThrows
   public static byte[] decryptAead(byte[] key, byte[] cipherText) {
-    if (key.length != 32) {
-      throw new AssertionError();
+    try {
+      if (key.length != 32) {
+        throw new AssertionError();
+      }
+      byte[] iv = ArrayUtils.subarray(cipherText, 0, GCM_IV_LENGTH);
+      byte[] ct = ArrayUtils.subarray(cipherText, GCM_IV_LENGTH, cipherText.length);
+
+      GCMParameterSpec ivParameterSpec = new GCMParameterSpec(GCM_TAG_LENGTH * Byte.SIZE, iv);
+      SecretKey secretKey = new SecretKeySpec(key, "AES");
+
+      Cipher cipher = Cipher.getInstance("AES/GCM/NoPadding");
+      cipher.init(Cipher.DECRYPT_MODE, secretKey, ivParameterSpec);
+      return cipher.doFinal(ct);
+    } catch (GeneralSecurityException e) {
+      throw new IllegalArgumentException("Error while decrypting AEAD ciphertext", e);
     }
-    byte[] iv = ArrayUtils.subarray(cipherText, 0, GCM_IV_LENGTH);
-    byte[] ct = ArrayUtils.subarray(cipherText, GCM_IV_LENGTH, cipherText.length);
-
-    GCMParameterSpec ivParameterSpec = new GCMParameterSpec(GCM_TAG_LENGTH * Byte.SIZE, iv);
-    SecretKey secretKey = new SecretKeySpec(key, "AES");
-
-    Cipher cipher = Cipher.getInstance("AES/GCM/NoPadding");
-    cipher.init(Cipher.DECRYPT_MODE, secretKey, ivParameterSpec);
-    return cipher.doFinal(ct);
   }
 
   /**
@@ -211,7 +220,6 @@ public class KEM {
    * @param encodedPublicKeyParameters contains key information
    * @return the reconstructed PublicKey
    */
-  @SneakyThrows
   public static ECPublicKey cborDecodeEcdhPublicKey(VauEccPublicKey encodedPublicKeyParameters) {
     return encodedPublicKeyParameters.toEcPublicKey();
   }