+++ src/main/java/de/gematik/vau/lib/data/VauBasicPublicKey.java
@@ -16,19 +16,25 @@
 
 package de.gematik.vau.lib.data;
 
+import static de.gematik.vau.lib.data.EccKyberKeyPair.KYBER_PUBLIC_KEY_ENCODING_HEADER;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import de.gematik.vau.lib.exceptions.VauKeyConversionException;
+import de.gematik.vau.lib.util.ArrayUtils;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.security.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.bouncycastle.asn1.ASN1InputStream;
+import org.bouncycastle.asn1.DERBitString;
+import org.bouncycastle.asn1.DLSequence;
 import org.bouncycastle.jce.interfaces.ECPublicKey;
 
-import java.security.KeyFactory;
-import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
-import java.security.PublicKey;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.X509EncodedKeySpec;
+import org.bouncycastle.util.encoders.Hex;
 
 @Data
 @AllArgsConstructor
@@ -43,11 +49,23 @@ public class VauBasicPublicKey {
   @SneakyThrows
   public VauBasicPublicKey(EccKyberKeyPair keyPair) {
     this.ecdhPublicKey = new VauEccPublicKey((ECPublicKey) keyPair.getEccKeyPair().getPublic());
-    this.kyberPublicKeyBytes = keyPair.getKyberKeyPair().getPublic().getEncoded();
+    this.kyberPublicKeyBytes = extractCompactKyberPublicKey(keyPair.getKyberKeyPair());
   }
 
-  public PublicKey toKyberPublicKey() throws NoSuchAlgorithmException, NoSuchProviderException, InvalidKeySpecException {
-    X509EncodedKeySpec x509EncodedKeySpec = new X509EncodedKeySpec(kyberPublicKeyBytes);
+  private byte[] extractCompactKyberPublicKey(KeyPair kyberKeyPair) {
+    try {
+      final byte[] verbosePublicKey = kyberKeyPair.getPublic().getEncoded();
+      final ASN1InputStream asn1InputStream = new ASN1InputStream(new ByteArrayInputStream(verbosePublicKey));
+      return ((DERBitString) ((DLSequence) asn1InputStream.readObject()).getObjectAt(1)).getBytes();
+    } catch (IOException e) {
+      throw new VauKeyConversionException("Error during key extraction for Kyber-key", e);
+    }
+  }
+
+  public PublicKey toKyberPublicKey()
+    throws NoSuchAlgorithmException, NoSuchProviderException, InvalidKeySpecException {
+    X509EncodedKeySpec x509EncodedKeySpec = new X509EncodedKeySpec(
+      ArrayUtils.unionByteArrays(KYBER_PUBLIC_KEY_ENCODING_HEADER, kyberPublicKeyBytes));
     KeyFactory keyFactory = KeyFactory.getInstance("KYBER", "BCPQC");
     return keyFactory.generatePublic(x509EncodedKeySpec);
   }